+++ JamGame/Assets/Scripts/Settings.meta
