+++ JamGame/Assets/Scripts/Settings/GlobalGameSettings.cs.meta
